+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - show username and discrim in addition to mention in `inrole` command
 - update to newest Elixir version
+- renamed USW "filters" to "rules"
 
 ## v0.7.0
 - add the `MENTIONS` filter